Move common/deletion_guard into util/dependency
